+++ AGENTS.md
@@ -2,12 +2,13 @@
 
 This project displays an interactive crossword in modern JavaScript and HTML.
 Puzzle data is loaded from `social_deduction_ok.xml` at runtime and rendered by
-`main.js` (an ES module). Open `index.html` in a browser to run the viewer.
+`index.js` (an ES module). Open `index.html` in a browser to run the viewer.
 
+Major modules: `crossword.js` implements the `Crossword` class and `puzzle-parser.js` handles XML parsing.
 ## Agent Tasks
 - Parse puzzle XML and build the grid and clues.
 - Maintain user input handling and diagnostic helpers.
-- Keep the code modular and readable. `main.js` exports a `crossword` instance
+- Keep the code modular and readable. `index.js` exports a `crossword` instance
   that exposes helper methods such as `testGridIsBuilt()` and
   `testCluesPresent()`.
 - Update this file when guidance changes and log notable updates in
@@ -17,7 +18,7 @@ Puzzle data is loaded from `social_deduction_ok.xml` at runtime and rendered by
 - Use modern ES6+ JavaScript and plain HTML/CSS.
 - No server-side code or external dependencies.
 - Add helpful console output for debugging. Enable verbose diagnostics by setting
-  `TEST_MODE` to `true` near the top of `main.js`.
+  `TEST_MODE` to `true` near the top of `index.js`.
 
 ## Design Notes
 - **DOM caching**: `buildGrid()` stores cell elements in
@@ -31,7 +32,7 @@ Puzzle data is loaded from `social_deduction_ok.xml` at runtime and rendered by
   clickable.
 - **Responsive sizing**: grid dimensions are calculated using the CSS variable
   `--cell-size` based on the viewport so the puzzle fits on mobile and desktop.
-- **Puzzle parsing**: `parsePuzzleData()` delegates to `parseGrid`, `parseClues`
+- **Puzzle parsing**: `parsePuzzle()` delegates to `parseGrid`, `parseClues`
   and `computeWordMetadata` helpers for clarity.
 
 ## Repository Practices
