+++ index.html
@@ -29,7 +29,7 @@
             <ul></ul>
         </div>
     </div>
-    <script type="module" src="main.js"></script>
+    <script type="module" src="index.js"></script>
 </body>
 </html>
